Crate astar_test_utils

Functions

call_wasm_contract_method

Call the wasm contract method and returns the decoded return values along with return flags and consumed weight

deploy_wasm_contract

and check for successful deploy

load_wasm_module

Load a given wasm module from wasm binary contents along with it’s hash.