Function astar_test_utils::call_wasm_contract_method

pub fn call_wasm_contract_method<T: Config, V: Decode>(
    origin: T::AccountId,
    dest: T::AccountId,
    value: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance,
    gas_limit: Weight,
    storage_deposit_limit: Option<<<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance>,
    data: Vec<u8>,
    debug: bool,
) -> (V, ReturnFlags, Weight)

Call the wasm contract method and returns the decoded return values along with return flags and consumed weight